smtp and sqlite to postgres

+++ comity/models.py
@@ -22,12 +22,12 @@ class group_info_table(models.Model):
     name = models.CharField(max_length=122)
     amount = models.CharField(max_length=6)
     duration = models.CharField(max_length=6, default='1m')
-    created_at=models.IntegerField(default=0)
-    updated_at=models.IntegerField(default=0)
-    start_date=models.IntegerField(default=0)
-    end_date=models.IntegerField(default=0)
+    created_at=models.BigIntegerField(default=0)
+    updated_at=models.BigIntegerField(default=0)
+    start_date=models.BigIntegerField(default=0)
+    end_date=models.BigIntegerField(default=0)
     status = models.CharField(max_length=12, default='inactive')
-    bid_date=models.IntegerField(default=0)
+    bid_date=models.BigIntegerField(default=0)
     created_by_user = models.ForeignKey(
         settings.AUTH_USER_MODEL,
         on_delete=models.CASCADE,
@@ -48,8 +48,8 @@ class UserInfo(models.Model):
     ifsc = models.CharField(max_length=20)
     branch_address = models.CharField(max_length=130, default = '')
     superuser = models.BooleanField(default=False)
-    superuser_start_date = models.IntegerField(default=0)
-    superuser_end_date = models.IntegerField(default=0)
+    superuser_start_date = models.BigIntegerField(default=0)
+    superuser_end_date = models.BigIntegerField(default=0)
     paytm_qr_code = models.ImageField(blank=True, upload_to='static/img/', default='')
     phonepe_qr_code = models.ImageField(blank=True, upload_to='static/img/', default='')
     gpay_qr_code = models.FileField(blank=True, upload_to='static/img/', default='')
